refactor: reduce duplicate code
